Actualizacion de Main
agrega faltantes metodos en el main General

+++ MainG.py
@@ -2,15 +2,25 @@ import os
 from DataBaseTree_AVL import AVLTree 
 import shutil
 
+
+
+
+
 class Main:
     
     def __init__(self):
         self.database= AVLTree()
         
-    
+    # CRUD DE LA BASE DE DATOS
     def dropDatabase(self,database: str) -> int: 
-        self.database.Eliminar(database)
-        shutil.rmtree("data/databases/"+database)
+        if self.verificador(database)==True:
+            self.database.Eliminar(database)
+            shutil.rmtree("data/databases/"+database)
+            return 0
+        elif self.verificador(database)==False:
+            return 2
+        else:
+            return 1
         
 
     def createDatabase(self,database:str)-> int:
@@ -43,6 +53,64 @@ class Main:
             return 1
 
 
+    #CRUD DE LAS TABLAS QUE ESTAN DENTRO DE LA BASE DE DATOS
+    def createTable(self,database: str, table: str, numberColumns: int) -> int:
+        print("CREAR TABLAS")
+
+    def alterAddPK(self,database: str, table:str , columns: list) -> int:
+        print("AGREGAR LLAVE PRIMARIA")
+
+    def alterDropPK(self,database: str, table:str) -> int:
+        print("ELIMINAR LLAVE PRIMARIA")
+
+    def defineFK(self,database: str, table:str, references: dict) -> int:
+        print("integridad de las tablas")
+
+    def showTables(self,database: str) -> list:
+        print("imprime las tablas en lista")
+
+    def alterTable(self,database: str, tableOld: str, tableNew: str) -> int:
+        print("Modifica el nombre de la tabla")
+
+    def dropTable(self,database: str, table:str) -> int:
+        print("eliminar tabla")
+
+    def alterAddColumn(self,database: str, table :str) -> int:
+        print("agregar columna")
+    
+    def alterDropColumn(self,database: str, table :str, columnNumber: int) -> int:
+        print("elimina columna")
+
+    
+    def extractTable(self,database: str, table :str) -> list:
+        print("extrae una  tabla y lo devuelve en lista")
+
+    
+    def extractRangeTable(self,database: str, table :str, lower: Any, upper: Any) -> list:
+        print("entrae una tabla y lo devuelve en lista pero en un rango especifico")
+
+    # FUNCIONES DENTRO DE LAS TUPLAS
+    def insert(self,database: str, table :str, register: list) -> int:
+        print("ingresa datos")
+
+    
+    def update(self,database: str, table :str, register: dict, columns: list) -> int:
+        print("inserta dato")
+
+    
+    def delete(self,database: str, table: str, columns: list) -> int:
+        print("elimina datos de una tabla")
+
+    
+    def truncate(self,database: str, table: str) -> int:
+        print("elimina un registro")
+
+    
+    def extractRow(self,database, table, id): 
+        print("extrae y devuelve una tupla especificada")
+
+
+    #METODOS EXTRAS 
     def initCheck(self,name):
         if not os.path.exists('data'):
             os.makedirs('data')
@@ -57,6 +125,12 @@ class Main:
             dum=True
         return dum
 
+
+
+
+
+
+
 s=Main()
 print(s.createDatabase("base1"))
 print(s.createDatabase("base2"))
@@ -66,5 +140,5 @@ print(s.createDatabase("base5"))
 print(s.createDatabase("base6"))
 print(s.createDatabase("base7"))
 print(s.dropDatabase("base3"))
-print(s.alterDatabase("base7","Holisbase7"))
+print(s.alterDatabase("Holisbase7","Holi"))
 s.showDatabases()
